release: merging to main is the release
The pipeline had grown a ceremony that existed entirely to serve npm. To cut a
release you ran release.ts, which bumped the root and three packages to one
version, committed, tagged, and told you to push. check-versions.ts then
enforced that alignment in CI, and promote depended on the npm job, so a
registry nobody had configured could hold `stable` hostage over a version
mismatch in a package no self-hoster installs.

Every failure of the first two real releases was in that job. The container
half was green throughout.

So npm moves to publish-npm.yml, manual, with its own version input. The two
are different products on different schedules: the image ships on every merge,
the packages ship when someone decides to cut them.

What is left is the part that was always working. Merging to main computes the
next patch from the newest tag, builds both architectures, merges the
manifest, proves the image installs and upgrades into the previous release,
then moves `stable`, tags the commit and writes the release notes. Nothing to
run locally and nothing to remember. Docs-only changes are ignored, because a
release costs two native builds and a full upgrade gate.

The tag is now written after the gate rather than by a human before it, so a
tag can no longer name a version that failed, and a failed run does not consume
a number. That also removes the lightweight-vs-annotated trap that silently
dropped v0.2.0 entirely.

Prerelease handling is gone with the manual tags that needed it. So is the
`latest` tag metadata-action was appending behind the gate's back: the tags
are now written directly, which is shorter than configuring the action not to.

Along the way, three things that blocked npm and would have blocked it again:
packages were never built before publishing, since the root build script is
`next build`; tsup's dts pass sets baseUrl internally, which TypeScript 6
rejects with TS5101; and isms-schema set baseUrl itself, needlessly, since its
paths are already relative to the file.

release.ts and check-versions.ts are deleted. set-package-versions.ts replaces
both with the only part still needed, and touches no git.
